# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POWERLVL v1 API contract — anonymous and member scans, public feed, leaderboards, profiles, and account.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{username}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a public Profile
      description: |
        Returns the public Profile header (username, avatar initials,
        aggregates) plus the first page of scan history (20 entries, newest
        first). 404 when the username does not match an existing Member.
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /profiles/{username}/scans:
    get:
      operationId: getProfileScans
      tags: [profiles]
      summary: Get a Profile's scan history page
      description: |
        Cursor-paginated scan history for a Member's public Profile, ordered
        by `createdAt DESC`.
      parameters:
        - $ref: "#/components/parameters/Username"
        - name: cursor
          in: query
          required: false
          description: ISO 8601 timestamp; entries strictly older than this are returned
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          description: Page size (default 20, max 20)
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 20
      responses:
        "200":
          description: Profile scan history page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileScansResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  parameters:
...
    Username:
      name: username
      in: path
      required: true
      description: Member username (lowercase, 3..20 chars)
      schema:
        type: string
        pattern: "^[a-z0-9_]{3,20}$"
        minLength: 3
        maxLength: 20
  responses:
...
    NotFound:
      description: Resource does not exist or has been expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorNotFound"
...
    Internal:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorInternal"
  schemas:
...
    Category:
      type: string
      description: Scan Category. Determines the five Category Stats.
      enum:
        - SETUPS
        - FITNESS
        - DRIP
        - PETS
        - RIDES
        - WILDCARD
    Tier:
      type: string
      description: Tier bucket derived from the final Score.
      enum:
        - D
        - C
        - B
        - A
        - S
        - SS
        - SSS
        - LIMITLESS
    AnomalyType:
      type: string
      description: Rare reveal event assigned to some Scans. At most one per Scan.
      enum:
        - POWER_SURGE_DETECTED
        - FORBIDDEN_AURA
        - SCOUTER_FAILURE
        - UNREGISTERED_ENERGY
        - CHAOS_SPIKE
...
    CoreStats:
      type: object
      description: The four Core Stats present on every Scan, each in the inclusive range 0..10000.
      additionalProperties: false
      properties:
        aura:
          type: integer
          minimum: 0
          maximum: 10000
        power:
          type: integer
          minimum: 0
          maximum: 10000
        status:
          type: integer
          minimum: 0
          maximum: 10000
        threat:
          type: integer
          minimum: 0
          maximum: 10000
      required: [aura, power, status, threat]
    CategoryStats:
      type: object
      description: |
        The five Category-specific stats produced for a Scan, each in the inclusive
        range 0..10000. Keys depend on the Scan's Category and are stable per
        Category (see design.md "Per-Category Stats"). Object has exactly five keys.
      additionalProperties:
        type: integer
        minimum: 0
        maximum: 10000
    ShareCardUrls:
      type: object
      description: Absolute URLs for the three share-card aspect ratios.
      additionalProperties: false
      properties:
        story:
          type: string
          format: uri
        square:
          type: string
          format: uri
        landscape:
          type: string
          format: uri
      required: [story, square, landscape]
    ScanPublic:
      type: object
      description: Public projection of a Scan, returned by Feed, Permalink, and Profile endpoints.
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        username:
          type: [string, "null"]
          description: Member username, or null when the Scan is unclaimed (anonymous).
          pattern: "^[a-z0-9_]{3,20}$"
        category:
          $ref: "#/components/schemas/Category"
        score:
          type: integer
          minimum: 1000
          maximum: 100000
        tier:
          $ref: "#/components/schemas/Tier"
        verdictNoun:
          type: string
          description: Verdict noun selected from the tier pool.
          minLength: 1
        coreStats:
          $ref: "#/components/schemas/CoreStats"
        categoryStats:
          $ref: "#/components/schemas/CategoryStats"
        commentary:
          type: string
          description: Final commentary line, 8..14 words inclusive.
          minLength: 1
        description:
          type: [string, "null"]
          description: Sanitized user-supplied description, ≤ 120 chars; null when absent.
          maxLength: 120
        anomaly:
          oneOf:
            - $ref: "#/components/schemas/AnomalyType"
            - type: "null"
        imageUrl:
          type: string
          format: uri
          description: Signed URL for the processed image (≤ 24 h validity).
        thumbnailUrl:
          type: string
          format: uri
          description: Signed URL for the 480px thumbnail (≤ 24 h validity).
        createdAt:
          type: string
          format: date-time
        permalink:
          type: string
          format: uri
          description: Absolute URL of the Scan permalink (`/scan/{id}`).
        shareCardUrls:
          $ref: "#/components/schemas/ShareCardUrls"
        likeCount:
          type: integer
          minimum: 0
        likedByMe:
          type: boolean
      required:
        - id
        - username
        - category
        - score
        - tier
        - verdictNoun
        - coreStats
        - categoryStats
        - commentary
        - description
        - anomaly
        - imageUrl
        - thumbnailUrl
        - createdAt
        - permalink
        - shareCardUrls
        - likeCount
        - likedByMe
...
    ProfileAggregates:
      type: object
      additionalProperties: false
      properties:
        highestScore:
          type: integer
          minimum: 0
          maximum: 100000
        totalScans:
          type: integer
          minimum: 0
        highestTier:
          $ref: "#/components/schemas/Tier"
      required: [highestScore, totalScans, highestTier]
    ProfilePublic:
      type: object
      description: Public Profile projection.
      additionalProperties: false
      properties:
        username:
          type: string
          pattern: "^[a-z0-9_]{3,20}$"
        avatarInitials:
          type: string
          minLength: 1
          maxLength: 2
          description: First letter of the username, uppercase.
        aggregates:
          $ref: "#/components/schemas/ProfileAggregates"
      required: [username, avatarInitials, aggregates]
    ProfileResponse:
      type: object
      description: Public Profile header plus the first page of scan history.
      additionalProperties: false
      properties:
        profile:
          $ref: "#/components/schemas/ProfilePublic"
        scans:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ScanPublic"
        nextCursor:
          type: [string, "null"]
          format: date-time
      required: [profile, scans, nextCursor]
    ProfileScansResponse:
      type: object
      additionalProperties: false
      properties:
        entries:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ScanPublic"
        nextCursor:
          type: [string, "null"]
          format: date-time
      required: [entries, nextCursor]
...
    ApiErrorNotFound:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [NOT_FOUND]
        message:
          type: string
      required: [code, message]
...
    ApiErrorInternal:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [INTERNAL]
        message:
          type: string
      required: [code, message]
...